added ability to track number of times movie was entered

+++ app.py
@@ -18,10 +18,12 @@ class Movie(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(120))
     year = db.Column(db.Integer)
+    entries = db.Column(db.Integer)
 
-    def __init__(self, title, year):
+    def __init__(self, title, year, entries):
         self.year = year
         self.title = title
+        self.entries = entries
     
     def __repr__(self):
         return '<Title %r>' % self.title
@@ -35,7 +37,7 @@ def movies_input():
     if request.content_type == 'application/json':
         post_data = request.get_json()
         user_title_input = post_data.get('title')
-        site = requests.get(f"https://www.imdb.com/find?ref_=nv_sr_fn&q={user_title_input.replace(' ', '+')}&s=all")
+        site = requests.get(f"https://www.imdb.com/find?q={user_title_input.replace(' ', '%20')}&s=tt&ttype=ft&ref_=fn_ft")
         soup = BS(site.content, 'html.parser')
         movie = soup.find_all('td', class_='result_text')[0]
         movie_name = movie.find('a').get_text()
@@ -45,27 +47,38 @@ def movies_input():
             verified_movie_year = movie_year_parsed[0]
         else:
             verified_movie_year = 0
-        reg = Movie(movie_name, verified_movie_year)
-        db.session.add(reg)
-        db.session.commit()
-        return jsonify('data posted')
-    return jsonify('')
+        reg = Movie(movie_name, verified_movie_year, 1)
+        check = Movie.query.filter_by(title=movie_name).count()
+        if check == 1:
+            movie_update = Movie.query.filter_by(title=movie_name).first()
+            movie_update.entries = movie_update.entries + 1
+            db.session.commit()
+            return jsonify('Movie already in database')
+        else: 
+            db.session.add(reg)
+            db.session.commit()
+            return jsonify('Movie Added')
+    return jsonify('Issue')
 
 @app.route('/movies_delete', methods=['DELETE'])
 def movies_remove():
     if request.content_type == 'application/json':
         post_data = request.get_json()
         movietitle = post_data.get('title')
-        obj=Movie.query.filter_by(title=movietitle).first()
-        db.session.delete(obj)
-        db.session.commit()
-        return jsonify('data deleted')
+        obj = Movie.query.filter_by(title=movietitle).first()
+        check = Movie.query.filter_by(title=movietitle).count()
+        if check == 1:
+            db.session.delete(obj)
+            db.session.commit()
+            return jsonify('Movie Deleted')
+        else: 
+            return jsonify('Movie Already Deleted')
     return jsonify('')
 
 @app.route('/return_movies', methods=['GET'])
 def movies_return():
     if request.method == 'GET':
-        all_movies = db.session.query(Movie.title, Movie.year).all()
+        all_movies = db.session.query(Movie.title, Movie.year, Movie.entries).all()
         return jsonify(all_movies)
 
 if __name__ == '__main__':
